Fixed explicit template instantiation errors for HostDeviceVector.
- replaced HostDeviceVector<unsigned int> with HostDeviceVector<int>

+++ src/common/host_device_vector.cc
@@ -153,9 +153,9 @@ void HostDeviceVector<T>::Reshard(GPUSet devices) const { }
 // explicit instantiations are required, as HostDeviceVector isn't header-only
 template class HostDeviceVector<bst_float>;
 template class HostDeviceVector<GradientPair>;
-template class HostDeviceVector<uint32_t>;
+template class HostDeviceVector<int>;
 template class HostDeviceVector<Entry>;
-template class HostDeviceVector<uint64_t>;
+template class HostDeviceVector<size_t>;
 
 }  // namespace xgboost
