+++ src/common/host_device_vector.cu
@@ -544,8 +544,8 @@ void HostDeviceVector<T>::Resize(size_t new_size, T v) {
 // explicit instantiations are required, as HostDeviceVector isn't header-only
 template class HostDeviceVector<bst_float>;
 template class HostDeviceVector<GradientPair>;
-template class HostDeviceVector<uint32_t>;
+template class HostDeviceVector<int>;
 template class HostDeviceVector<Entry>;
-template class HostDeviceVector<uint64_t>;
+template class HostDeviceVector<size_t>;
 
 }  // namespace xgboost
